Added: option for normalising playback loudness, keybinding for grid sizing, clustering presets and output device chooser. Fixed issues with some wav files. Instant reset of library synch info window. Calculating absolute dynamic range instead of min/max values.

+++ Saempl/Source/SampleFileFilterRuleLength.cpp
@@ -24,7 +24,7 @@ SampleFileFilterRuleLength::~SampleFileFilterRuleLength()
 
 bool SampleFileFilterRuleLength::matches(SampleItem const & inSampleItem)
 {
-    int propertyValue = inSampleItem.getLength();
+    double propertyValue = inSampleItem.getLength();
     
     switch (mCompareOperator)
     {
